% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feeling_chicken.R
\name{feeling_chicken}
\alias{feeling_chicken}
\title{Feeling Chicken? This will uplift you!}
\usage{
feeling_chicken()
}
\description{
Feeling Chicken? This will uplift you!
}
\examples{
feeling_chicken()

}
